+++ src/main/java/org/folio/oaipmh/mappers/MarcXmlMapper.java
@@ -6,9 +6,11 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.UncheckedIOException;
 import java.nio.charset.StandardCharsets;
+import java.util.UUID;
 import java.util.regex.Pattern;
 
 import org.apache.commons.lang3.time.StopWatch;
+import org.folio.oaipmh.service.MetricsCollectingService;
 import org.marc4j.MarcJsonReader;
 import org.marc4j.MarcReader;
 import org.marc4j.MarcXmlWriter;
@@ -16,6 +18,8 @@ import org.marc4j.marc.Record;
 
 import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.LogManager;
+
+import static org.folio.oaipmh.service.MetricsCollectingService.MetricOperation.PARSE_XML;
 
 /**
  * Converts MarcJson format to MarcXML format.
@@ -26,6 +30,8 @@ public class MarcXmlMapper implements Mapper {
 
   private static final Pattern DOUBLE_BACKSLASH_PATTERN = Pattern.compile("\\\\\\\\");
 
+  private MetricsCollectingService metricsCollectingService = MetricsCollectingService.getInstance();
+
   /**
    * Convert MarcJson to MarcXML.
    *
@@ -33,6 +39,10 @@ public class MarcXmlMapper implements Mapper {
    * @return byte[] representation of MarcXML
    */
   public byte[] convert(String source) {
+
+    var operationId = UUID.randomUUID().toString();
+    metricsCollectingService.startMetric(operationId, PARSE_XML);
+
     StopWatch timer = logger.isDebugEnabled() ? StopWatch.createStarted() : null;
     /*
      * Fix indicators which comes like "ind1": "\\" in the source string and values are converted to '\'
@@ -49,11 +59,11 @@ public class MarcXmlMapper implements Mapper {
     } catch (IOException e) {
       throw new UncheckedIOException(e);
     } finally {
+      metricsCollectingService.endMetric(operationId, PARSE_XML);
       if (timer != null) {
         timer.stop();
         logger.debug("Marc-json converted to MarcXml after {} ms.", timer.getTime());
       }
     }
   }
-
 }