+++ src/test/java/org/folio/rest/impl/OkapiMockServer.java
@@ -10,6 +10,7 @@ import static org.folio.oaipmh.Constants.LOCATION_URI;
 import static org.folio.oaipmh.Constants.MATERIAL_TYPES_URI;
 import static org.folio.oaipmh.Constants.OKAPI_TENANT;
 import static org.folio.oaipmh.Constants.RESOURCE_TYPES_URI;
+import static org.folio.oaipmh.processors.MarcWithHoldingsRequestHelper.SKIP_SUPPRESSED_FROM_DISCOVERY_RECORDS;
 import static org.junit.jupiter.api.Assertions.fail;
 
 import java.io.File;
@@ -99,6 +100,11 @@ public class OkapiMockServer {
   static final String SRS_RECORDS_WITH_CYRILLIC_DATA_DATE = "2002-02-02";
   static final String SUPPRESSED_RECORDS_DATE = "2020-03-30";
   static final String NO_ITEMS_DATE = "2020-01-29";
+  static final String GET_INSTANCES_FORBIDDEN_RESPONSE_DATE = "2020-01-10";
+  static final String ENRICH_INSTANCES_FORBIDDEN_RESPONSE_DATE = "2020-01-11";
+  static final String GET_INSTANCES_IDS_500_ERROR_RETURNED_FROM_STORAGE_DATE = "2020-01-12";
+  static final String GET_ENRICHED_INSTANCES_500_ERROR_RETURNED_FROM_STORAGE_DATE = "2020-01-12";
+
   public static final String INVALID_INSTANCE_IDS_JSON_DATE = "2011-11-22";
   public static final String INSTANCE_ID_WITH_INVALID_ENRICHED_INSTANCE_JSON_DATE = "2012-11-22";
 
@@ -174,14 +180,21 @@ public class OkapiMockServer {
   private static final String INVALID_JSON = "invalid.json";
   private static final String INSTANCE_ID_INVALID_ENRICHED_INSTANCE_JSON = "instance_id_invalid_enriched_instance.json";
   private static final String INSTANCE_IDS_UNDERLYING_SRS_RECORDS_WITH_CYRILLIC_JSON = "instance_ids_underlying_srs_records_with_cyrillic.json";
+  private static final String INSTANCE_ID_ENRICH_INSTANCES_FORBIDDEN_RESPONSE_JSON = "inventory_instance_mock_forbidden_response.json";
+  private static final String INSTANCE_ID_ENRICH_INSTANCES_500_RESPONSE_JSON = "inventory_instance_mock_500_response.json";
+
   private static final String INSTANCE_ID_NO_SRS_RECORD_JSON = "instance_id_no_srs_record.json";
   private static final String INSTANCE_ID_UNDERLYING_RECORD_WITH_CYRILLIC_DATA = "ebbb759a-dd08-4bf8-b3c3-3d75b2190c41";
   private static final String INSTANCE_ID_WITHOUT_SRS_RECORD = "3a6a47ab-597d-4abe-916d-e31c723426d3";
+  private static final String INSTANCE_ID_ENRICH_INSTANCES_FORBIDDEN_RESPONSE = "8f33cdf4-6a85-4877-8b99-7d5e3be910f1";
+  private static final String INSTANCE_ID_ENRICH_INSTANCES_500_RESPONSE = "12d31a35-e6cf-4840-bd85-4ae51e02a741";
+
 
   private static final String JSON_TEMPLATE_KEY_RECORDS = "replace_with_records";
   private static final String JSON_TEMPLATE_KEY_TOTAL_COUNT = "replace_total_count";
   private static final int LAST_RECORDS_BATCH_OFFSET_VALUE = 8;
   private static final int LIMIT_VALUE_FOR_LAST_TWO_RECORDS_IN_JSON = 2;
+  private static final String INTERNAL_SERVER_ERROR = "Internal Server Error";
 
   private static int srsRerequestAttemptsCount = 4;
   private static int totalSrsRerequestCallsNumber = 0;
@@ -245,7 +258,7 @@ public class OkapiMockServer {
       .absoluteURI();
     if (Objects.nonNull(uri)) {
       if (uri.contains(SUPPRESSED_RECORDS_DATE)) {
-        boolean shouldProcessSuppressedRecords = Boolean.parseBoolean(ctx.request().getParam("skipSuppressedFromDiscoveryRecords"));
+        boolean shouldProcessSuppressedRecords = Boolean.parseBoolean(ctx.request().getParam(SKIP_SUPPRESSED_FROM_DISCOVERY_RECORDS));
         if (shouldProcessSuppressedRecords) {
           inventoryViewSuccessResponse(ctx, INSTANCE_IDS_10_JSON);
         } else {
@@ -253,7 +266,7 @@ public class OkapiMockServer {
         }
       }
       if (uri.contains(DATE_INVENTORY_STORAGE_ERROR_RESPONSE)) {
-        failureResponse(ctx);
+        failureResponseWithForbidden(ctx);
       } else if (uri.contains(DATE_INVENTORY_10_INSTANCE_IDS)) {
         inventoryViewSuccessResponse(ctx, INSTANCE_IDS_10_JSON);
       } else if (uri.contains(INVENTORY_27_INSTANCES_IDS_DATE)) {
@@ -284,6 +297,14 @@ public class OkapiMockServer {
         inventoryViewSuccessResponse(ctx, INSTANCE_ID_NO_SRS_RECORD_JSON);
       } else if (uri.contains(NO_ITEMS_DATE)) {
         inventoryViewSuccessResponse(ctx, ENRICHED_INSTANCE_NO_ITEMS_JSON);
+      } else if (uri.contains(GET_INSTANCES_FORBIDDEN_RESPONSE_DATE)) {
+        failureResponseWithForbidden(ctx);
+      } else if (uri.contains(GET_INSTANCES_IDS_500_ERROR_RETURNED_FROM_STORAGE_DATE)) {
+        failureResponse(ctx, 500, INTERNAL_SERVER_ERROR);
+      } else if (uri.contains(GET_ENRICHED_INSTANCES_500_ERROR_RETURNED_FROM_STORAGE_DATE)) {
+        inventoryViewSuccessResponse(ctx, INSTANCE_ID_ENRICH_INSTANCES_500_RESPONSE_JSON);
+      } else if (uri.contains(ENRICH_INSTANCES_FORBIDDEN_RESPONSE_DATE)) {
+        inventoryViewSuccessResponse(ctx, INSTANCE_ID_ENRICH_INSTANCES_FORBIDDEN_RESPONSE_JSON);
       } else {
         logger.debug("No mocks for the response, returning the default instance id.");
         inventoryViewSuccessResponse(ctx, DEFAULT_INSTANCE_JSON);
@@ -298,13 +319,17 @@ public class OkapiMockServer {
       .getJsonArray(INSTANCE_IDS);
     logger.debug("Before building response for enriched instances, instanceIds: {}.", String.join(",", instanceIds.getList()));
     if (instanceIds.contains(INSTANCE_ID_TO_FAIL_ENRICHED_INSTANCES_REQUEST)) {
-      failureResponse(ctx);
+      failureResponseWithForbidden(ctx);
     } else if (instanceIds.contains(INSTANCE_ID_RELATED_ENRICHED_INSTANCE_HAS_INVALID_JSON)) {
       successResponse(ctx, getJsonObjectFromFileAsString(INVENTORY_VIEW_PATH + INVALID_JSON));
     } else if (instanceIds.isEmpty()) {
       successResponse(ctx, "");
     } else if (instanceIds.contains(INSTANCE_ID_RELATED_ENRICHED_INSTANCE_HAS_NO_ITEMS)) {
       inventoryViewSuccessResponse(ctx, ENRICHED_INSTANCE_NO_ITEMS_JSON);
+    } else if (instanceIds.contains(INSTANCE_ID_ENRICH_INSTANCES_FORBIDDEN_RESPONSE)) {
+      failureResponseWithForbidden(ctx);
+    } else if (instanceIds.contains(INSTANCE_ID_ENRICH_INSTANCES_500_RESPONSE)) {
+      failureResponse(ctx, 500, INTERNAL_SERVER_ERROR);
     } else {
       inventoryViewSuccessResponse(ctx, instanceIds);
     }
@@ -369,7 +394,7 @@ public class OkapiMockServer {
     JsonArray instanceIds = ctx.getBody()
       .toJsonArray();
     if (instanceIds.contains(INSTANCE_ID_TO_MAKE_SRS_FAIL)) {
-      failureResponse(ctx);
+      failureResponseWithForbidden(ctx);
     } else if (instanceIds.contains(INSTANCE_ID_TO_MAKE_SRS_FAIL_WITH_500)) {
       totalSrsRerequestCallsNumber++;
       if (srsRerequestAttemptsCount > 0) {
@@ -531,7 +556,7 @@ public class OkapiMockServer {
       .end(buffer);
   }
 
-  private void failureResponse(RoutingContext ctx) {
+  private void failureResponseWithForbidden(RoutingContext ctx) {
     ctx.response()
       .setStatusCode(403)
       .setStatusMessage("Forbidden")